refactored VISTA VSA + UKIDSS to use a common WFAU tool
